pyproject.toml
[project]
name = "pipeline-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.179.0",
    "constructs>=10.0.0,<11.0.0",
]

[dependency-groups]
dev = [
    "aws-lambda-powertools>=3.6.0",
    "pytest==6.2.5",
]
